calculator_backend/requirements.txt
anyio==4.0.0
asgiref==3.7.2
astor==0.8.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
decorator==5.1.1
distlib==0.3.6
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.1.3
filelock==3.9.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-vision==3.4.5
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
numpy==1.26.1
opencv-contrib-python==4.8.1.78
opencv-contrib-python-headless==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
paddlepaddle==2.5.2
Pillow==10.1.0
platformdirs==3.0.0
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytesseract==0.3.10
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.19.0
